requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
anykeystore==0.2
apex==0.9.10.dev0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autocommand==2.2.2
Babel==2.14.0
backports.csv==1.0.7
backports.tarfile==1.1.1
beautifulsoup4==4.12.3
bert-score==0.3.13
bitsandbytes==0.41.0
black==24.2.0
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
blobfile==2.1.1
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cheroot==10.0.1
CherryPy==18.9.0
click==8.1.7
cliffs-delta==1.0.0
cloudpathlib==0.16.0
cmake==3.28.3
colorama==0.4.6
comm==0.2.1
confection==0.1.4
configparser==7.1.0
contourpy==1.2.0
cryptacular==1.6.2
cryptography==42.0.5
cupy-cuda12x==13.0.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
embeddings==0.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.13
faiss-cpu==1.7.4
fastapi==0.109.2
fastjsonschema==2.19.1
fastrlock==0.8.2
feedparser==6.0.11
ffmpy==0.3.2
filelock==3.13.1
flake8==7.0.0
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
future==1.0.0
fuzzysearch==0.7.3
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.42
gradio==4.16.0
gradio_client==0.8.1
greenlet==3.0.3
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.20.3
hupper==1.12.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflect==7.2.1
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.2
ipython==8.18.1
isoduration==20.11.0
jaraco.collections==5.0.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.0
jdk4py==21.0.2.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
kiwisolver==1.4.5
langcodes==3.3.0
lightning==2.2.0.post0
lightning-utilities==0.10.1
linkify-it-py==2.0.3
lit==17.0.6
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
lxml==4.9.4
markdown-it-py==2.2.0
markdown2==2.4.12
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mauve-text==0.3.0
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mpire==2.10.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy==1.8.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.12.0
openai-clip==1.0.1
opencv-python-headless==4.9.0.80
orjson==3.9.14
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
PasteDeploy==3.1.0
pathspec==0.12.1
Pattern==3.6
pbkdf2==1.3
pdfminer.six==20231228
peft==0.4.0
pexpect==4.9.0
pillow==10.2.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.8.2
portend==3.2.0
POT==0.9.3
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.17.2
PyMySQL==1.1.1
pyparsing==3.1.1
pyramid==2.0.2
pyramid-mailer==0.15.1
pytest==8.0.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python3-openid==3.2.0
pytorch-lightning==2.2.0.post0
pytorch-minimize==0.0.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rouge-score==0.1.2
rpds-py==0.18.0
ruff==0.4.5
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.3.1
sentencepiece==0.1.99
sentry-sdk==2.1.1
setproctitle==1.3.3
sgmllib3k==1.0.0
sh==2.0.6
shellingham==1.5.4
shortuuid==1.0.11
shutils==0.1.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
-e git+https://github.com/spetryk/vl-hallucination.git@4ee0b5bf3246a98c3e73160fa2ee1bc0ec4586d0#egg=src
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.3
streamlit==1.31.1
svgwrite==1.4.3
sympy==1.12
tempora==5.5.1
tenacity==8.2.3
tensorboardX==2.6.2.2
terminado==0.18.0
thinc==8.2.3
threadpoolctl==3.3.0
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchmetrics==1.3.1
torchvision==0.16.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transaction==4.0
transformers==4.37.2
translationstring==1.4
triton==2.1.0
typeguard==4.2.1
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.27.1
validators==0.22.0
vdtk==0.3.0
velruse==1.1.1
venusian==3.1.0
wandb==0.17.0
wasabi==1.1.2
watchdog==4.0.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webdataset==0.2.86
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.7.0
websockets==11.0.3
WTForms==3.1.2
wtforms-recaptcha==0.3.2
xformers==0.0.22
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
zope.deprecation==5.0
zope.interface==6.4
zope.sqlalchemy==3.1
